ROVER-244 Deal with flags properly
